refactor: 요구사항 수정
