+++ src/app/(with-protected)/my-quotes/page.tsx
@@ -1,17 +1,6 @@
-"use client";
+import MyQuotes from "@/components/my-quotes/MyQuotes";
 
-import Pending from "@/components/my-quotes/pages/Pending";
-import Received from "@/components/my-quotes/pages/Received";
-import { useMyQuotes } from "@/context/MyQuotesContext";
-
+// 내 견적 관리
 export default function Page() {
-  const { activeTab } = useMyQuotes();
-
-  if (activeTab === "pending") {
-    return <Pending />;
-  } else if (activeTab === "received") {
-    return <Received />;
-  }
-
-  return null;
+  return <MyQuotes />;
 }
