+++ src/components/my-quotes/common/MoveDateCard.tsx
@@ -1,11 +1,9 @@
-import React from "react";
-
 interface MoveDateCardProps {
   category: string;
   text: string;
 }
 
-const MoveDateCard = ({ category, text }: MoveDateCardProps) => {
+export default function MoveDateCard({ category, text }: MoveDateCardProps) {
   return (
     <div className="flex items-center gap-2">
       <div className="px-1.5 bg-bg-400 rounded">
@@ -14,6 +12,4 @@ const MoveDateCard = ({ category, text }: MoveDateCardProps) => {
       <p className="text-14-medium">{text}</p>
     </div>
   );
-};
-
-export default MoveDateCard;
+}
