+++ src/context/MyQuotesContext.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { createContext, ReactNode, useContext, useState } from "react";
+import { createContext, ReactNode, useState } from "react";
 
 type Tab = "pending" | "received";
 
@@ -9,8 +9,11 @@ interface MyQuotesContext {
   setActiveTab: (tab: Tab) => void;
 }
 
-const MyQuotesContext = createContext<MyQuotesContext | undefined>(undefined);
+export const MyQuotesContext = createContext<MyQuotesContext | undefined>(
+  undefined
+);
 
+// 헤더에서 컴포넌트로 변경을 위한 provider
 export function MyQuotesProvider({ children }: { children: ReactNode }) {
   const [activeTab, setActiveTab] = useState<Tab>("pending");
 
@@ -21,10 +24,3 @@ export function MyQuotesProvider({ children }: { children: ReactNode }) {
   );
 }
 
-export const useMyQuotes = () => {
-  const context = useContext(MyQuotesContext);
-  if (context === undefined) {
-    throw new Error("useMyQuotes must be used within a MyQuotesProvider");
-  }
-  return context;
-};
