+++ src/lib/hooks/useMyQuotes.tsx
@@ -1,0 +1,12 @@
+"use client";
+
+import { MyQuotesContext } from "@/context/MyQuotesContext";
+import { useContext } from "react";
+
+export const useMyQuotes = () => {
+  const context = useContext(MyQuotesContext);
+  if (context === undefined) {
+    throw new Error("useMyQuotes must be used within a MyQuotesProvider");
+  }
+  return context;
+};
